Move pytest configuration to shared conftest
