+++ ci/testkomodo.sh
@@ -36,8 +36,7 @@ run_ert_with_opm() {
 
 run_everest_tests() {
     python -m pytest tests/everest -s \
-    --ignore-glob "*test_visualization_entry*" \
-     -m "not requires_eclipse"
+    --ignore-glob "*test_visualization_entry*"
 
 }
