+++ tests/ert/conftest.py
@@ -265,37 +265,6 @@ def hide_window(request):
         os.environ["QT_QPA_PLATFORM"] = old_value
 
 
-def pytest_addoption(parser):
-    parser.addoption(
-        "--runslow", action="store_true", default=False, help="run slow tests"
-    )
-    parser.addoption(
-        "--eclipse-simulator",
-        action="store_true",
-        default=False,
-        help="Defaults to not running tests that require eclipse.",
-    )
-    parser.addoption(
-        "--openpbs",
-        action="store_true",
-        default=False,
-        help="Run OpenPBS tests against the real cluster",
-    )
-    parser.addoption(
-        "--lsf",
-        action="store_true",
-        default=False,
-        help="Run LSF tests against the real cluster.",
-    )
-    parser.addoption(
-        "--slurm",
-        action="store_true",
-        default=False,
-        help="Run Slurm tests against a real cluster.",
-    )
-    parser.addoption("--show-gui", action="store_true", default=False)
-
-
 @pytest.fixture
 def _qt_excepthook(monkeypatch):
     """Hook into Python's unhandled exception handler and quit Qt if it's
@@ -311,37 +280,6 @@ def _qt_excepthook(monkeypatch):
         next_excepthook(cls, exc, tb)
 
     monkeypatch.setattr(sys, "excepthook", excepthook)
-
-
-def pytest_collection_modifyitems(config, items):
-    for item in items:
-        fixtures = getattr(item, "fixturenames", ())
-        if "qtbot" in fixtures or "qapp" in fixtures or "qtmodeltester" in fixtures:
-            item.add_marker("requires_window_manager")
-
-    # Override Python's excepthook on all "requires_window_manager" tests
-    for item in items:
-        if item.get_closest_marker("requires_window_manager"):
-            item.fixturenames.append("_qt_excepthook")
-            item.fixturenames.append("_qt_add_search_paths")
-        if item.get_closest_marker("requires_eclipse") and not config.getoption(
-            "--eclipse-simulator"
-        ):
-            item.add_marker(pytest.mark.skip("Requires eclipse"))
-
-    if config.getoption("--runslow"):
-        # --runslow given in cli: do not skip slow tests
-        skip_quick = pytest.mark.skip(
-            reason="skipping quick performance tests on --runslow"
-        )
-        for item in items:
-            if "quick_only" in item.keywords:
-                item.add_marker(skip_quick)
-    else:
-        skip_slow = pytest.mark.skip(reason="need --runslow option to run")
-        for item in items:
-            if "slow" in item.keywords:
-                item.add_marker(skip_slow)
 
 
 def _run_snake_oil(source_root):
